--- config/test_dtm0_config.yaml
@@ -1,20 +1,0 @@
-test_41204:
-  clients: 1
-  samples: 500
-  size: '128Mb'
-  log_file_prefix: 'test-40039'
-test_41219:
-  clients: 1
-  samples: 500
-  size: '128Mb'
-  log_file_prefix: 'test-41219'
-test_41223:
-  clients: 1
-  samples: 500
-  size: '128Mb'
-  log_file_prefix: 'test-41223'
-test_41224:
-  clients: 1
-  samples: 500
-  size: '128Mb'
-  log_file_prefix: 'test-41224'
